moved endpoints to internal
