+++ internal/endpoints/endpoint.go
@@ -1,18 +1,20 @@
-package advisor
+package endpoints
 
 import (
 	"net/http"
 	"strconv"
 
+	"github.com/romietis/lunar-plan-advisor/advisor"
+
 	"github.com/gin-gonic/gin"
 )
 
-func Endpoint(c *gin.Context) {
+func Home(c *gin.Context) {
 	c.HTML(http.StatusOK, "index.html", nil)
 }
 
-func PlansEndpoint(c *gin.Context) {
-	planConfig := []Plan{
+func GetPlans(c *gin.Context) {
+	planConfig := []advisor.Plan{
 		{Name: "Light", AnnualInterestRate: 1.5, Fee: 0.0, Cap: 100000},
 		{Name: "Standard", AnnualInterestRate: 1.75, Fee: 29.0, Cap: 100000},
 		{Name: "Plus", AnnualInterestRate: 2.0, Fee: 69.0, Cap: 0},
@@ -35,7 +37,7 @@ func PlansEndpoint(c *gin.Context) {
 		return
 	}
 
-	bestPlans, err := CalculatePlans(balance_float, planConfig)
+	bestPlans, err := advisor.CalculatePlans(balance_float, planConfig)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": err.Error(),
